fix(sec): satisfy the clang-tidy families REQ-SEC-015 makes fatal
First real clang-tidy run (OQ-026) surfaced the bugprone/cert findings the
.clang-tidy WarningsAsErrors names as errors, all in pre-existing code:

- two duplicated switch branches in strip_diacritic (C6/E6 folded into
  their groups) and misplaced widening casts in parse_hex — widen the
  operand, then subtract
- the 8 MiB limit in json.hpp computed in 32-bit unsigned before widening,
  and a same-shape cast in dump_into
- main() let exceptions escape; the body is now wrapped so an unexpected
  throw exits EX_SOFTWARE instead of std::terminate, and the error-code
  string_view is printed with an explicit length rather than assumed null
  termination
- ignored fprintf/fflush returns across main and the fuzz harnesses are
  marked as deliberately discarded, and the two deterministic test RNGs
  carry NOLINT explaining that reproducibility is the point

Refs: REQ-SEC-015

+++ desktop/src/core/json/json.cpp
@@ -194,7 +194,7 @@ void dump_into(const Value& v, int indent, int level, std::string& out) {
     const auto newline_indent = [&](int lvl) {
         if (indent <= 0) return;
         out.push_back('\n');
-        out.append(static_cast<std::size_t>(indent * lvl), ' ');
+        out.append(static_cast<std::size_t>(indent) * static_cast<std::size_t>(lvl), ' ');
     };
 
     switch (v.type()) {
